I added borders to each individual grid.
I needed the borders in order to see if the grid was working properly (for debugging), but also I believe the borders make the grid more organized and clean for the user to interact. The issue is that the borders make the grids look ugly when you have painted them. To counteract this, I plan to delete the border when grids are painted and give it back when grids are erased (painted white).